feat(nvim,lsp,rust): add command for setting cargo target
I need to figure out how to do this without defer, or maybe even
reloading. The workspace seems to need to be reloaded after the
notification due to the [inactive-code] diagnostic, but the reload
request being processed before the notification is completed, so I can
only get it to work by defering. This technically introduces a race
condition?